adds salutation column to schema and entity
