+++ src/main/java/com/otsMail/util/AppConstants.java
@@ -3,5 +3,6 @@ package com.otsMail.util;
 public class AppConstants {
 	public static final String API = "/api/v1";
 	public static final Integer DEFAULT_COUNT = 0;
+	public static final String DefaultSalutaion = "Team";
 
 }
