+++ src/main/resources/schema.sql
@@ -2,6 +2,7 @@ CREATE TABLE "EmailTrack" (
     "Id" BIGINT PRIMARY KEY AUTO_INCREMENT,
     "Sender" VARCHAR(255) NOT NULL,
     "Receipient" VARCHAR(255) NOT NULL,
+    "Salutation" VARCHAR(255),
     "Time" TIMESTAMP NOT NULL,
     "Status" VARCHAR(255) NOT NULL,
     "Count" INT,
